clean up topic dropdown
